backend/requirements.txt
# FastAPI Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.23.0

# Database ORM
sqlmodel>=0.0.14
asyncpg>=0.29.0

# Authentication & JWT
python-jose[cryptography]>=3.3.0
httpx>=0.25.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# File Upload Support
python-multipart>=0.0.6

# Development Dependencies
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
ruff>=0.1.0
